feat: consider params in request caching
This commit introduces (naive) parameter consideration to request
caching - meaning that parameters will be considered when deciding
whether to use an existing cached value.

The approach is naive in that it does not consider defaults, so
additional work might be performed if a user requests something w/ the
same parameters as the defaults. That being said, in a "real-world"
workload, this seems extremely unlikely.